djangooauth/requirements.txt
Django==5.1.6
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.32.3
cryptography==42.0.5